@@ -44,7 +44,7 @@ const ComicsList = () => {
                     <a href="#">
                         <img src={item.thumbnail} alt={item.name} className="comics__item-img"/>
                         <div className="comics__item-name">{item.name}</div>
-                        <div className="comics__item-price">${item.price}</div>
+                        <div className="comics__item-price">{item.price}</div>
                     </a>
                 </li>
             );
